refactor(rust): extract initiator and responder code into own modules
