+++ implementations/rust/daemon/src/cli.rs
@@ -175,7 +175,7 @@ impl FromStr for VaultKind {
 }
 
 /// Specifies which end of the secure channel the instance of `ockamd` is prepared to run in.
-#[derive(Clone, Copy, Debug, StructOpt, PartialEq)]
+#[derive(Clone, Copy, Debug, StructOpt)]
 pub enum ChannelRole {
     /// The Initiator role expects a channel responder address and a public key to use in order to
     /// communicate with the Responder end of the channel.
