+++ implementations/rust/daemon/src/lib.rs
@@ -1,4 +1,6 @@
 pub mod cli;
+pub mod initiator;
 pub mod node;
+pub mod responder;
 pub mod vault;
 pub mod worker;
